tgbot_project/requirements.txt
python-telegram-bot==13.15
tweepy>=4.14,<5
SQLAlchemy>=1.4,<3
PyYAML>=6.0
